+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.6
 ----------------
 - Fix #157 Delete Period when days unchecked in Courses.php
+- Fix Check subject ID is valid for current school & syear in Courses.php
 
 Changes in 2.9.5
 ----------------